Separate liballoc module
